Add settings for logical address and device type.
Auto-negotiation of all HDMI-CEC parameters is usually reliable, but not
always correct or desired.
Similar to setting physical address, add support for changing the device
type and hard coding the logical address.

The logical address dictates the device type but no checks are performed
to ensure consistency.
The device type determines which set of logical addresses will attempt to
be allocated.

Add more CEC message decoding for logging.

Further modifications to routing control to better comply with the HDMI
1.3a specification.